Refactor: A trait to get cast crew details
Created a trait so as to avoid the code duplication for cast and crew
details, on the movie details page and on the TV show details page.
